+++ flutter-hms-gameservice/android/src/main/java/com/huawei/hms/flutter/gameservice/HuaweiGameServicePlugin.java
@@ -23,10 +23,8 @@ import androidx.annotation.NonNull;
 import com.huawei.hms.api.HuaweiMobileServicesUtil;
 import com.huawei.hms.flutter.gameservice.common.Constants;
 import com.huawei.hms.flutter.gameservice.controllers.GameServiceController;
-import com.huawei.hms.jos.JosApps;
-import com.huawei.hms.jos.JosAppsClient;
+import com.huawei.hms.flutter.gameservice.controllers.JosAppsClientController;
 
-import io.flutter.Log;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
@@ -34,38 +32,39 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
 public class HuaweiGameServicePlugin implements FlutterPlugin, ActivityAware {
-    private static final String TAG = "HuaweiGameServicePlugin";
 
     private Activity activity;
 
     private MethodChannel gameServiceChannel;
 
+    private MethodChannel josAppsChannel;
+
     public static void registerWith(Registrar registrar) {
         HuaweiMobileServicesUtil.setApplication(registrar.activity().getApplication());
-        JosAppsClient appsClient = JosApps.getJosAppsClient(registrar.activity());
-        appsClient.init();
         final MethodChannel gameServiceChannel = new MethodChannel(registrar.messenger(),
             Constants.Channel.GAME_SERVICE_METHOD_CHANNEL);
         gameServiceChannel.setMethodCallHandler(new GameServiceController(registrar.activity(), gameServiceChannel));
-    }
-
-    private void init() {
-        JosAppsClient appsClient = JosApps.getJosAppsClient(activity);
-        appsClient.init();
-        Log.i(TAG, "init success");
+        final MethodChannel josAppsChannel = new MethodChannel(registrar.messenger(),
+            Constants.Channel.JOS_APPS_METHOD_CHANNEL);
+        josAppsChannel.setMethodCallHandler(new JosAppsClientController(registrar.activity(), josAppsChannel));
     }
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
         gameServiceChannel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(),
             Constants.Channel.GAME_SERVICE_METHOD_CHANNEL);
+        josAppsChannel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(),
+            Constants.Channel.JOS_APPS_METHOD_CHANNEL);
     }
 
     @Override
     public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
         if (gameServiceChannel != null) {
-            gameServiceChannel.setMethodCallHandler(null);
             teardownMethodChannel(gameServiceChannel);
+        }
+
+        if (josAppsChannel != null) {
+            teardownMethodChannel(josAppsChannel);
         }
     }
 
@@ -73,9 +72,12 @@ public class HuaweiGameServicePlugin implements FlutterPlugin, ActivityAware {
     public void onAttachedToActivity(ActivityPluginBinding binding) {
         this.activity = binding.getActivity();
         HuaweiMobileServicesUtil.setApplication(activity.getApplication());
-        init();
         if (gameServiceChannel != null) {
             gameServiceChannel.setMethodCallHandler(new GameServiceController(activity, gameServiceChannel));
+        }
+
+        if (josAppsChannel != null) {
+            josAppsChannel.setMethodCallHandler(new JosAppsClientController(activity, josAppsChannel));
         }
     }